refactor: Extract EditorjsTextBlock component and clean up CustomFile and ImageBlock
